+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5.4 (WIP)
 
 * Set a `timeout` and `proxyTimeout` in `SiteDispatcher#forwardRequest()`
+* Do not retry connection on error
 
 ## 0.5.3 (2023-10-19)